directionsToGoForEvilUsingMouse is debugged
Since we were only using X and Y coordinates of the evil circle, hitbox of the circle was rectangular instead of circular, resulting in movement stopping early.
To solve this, when the mouse is inside X and Y coordinates, we are also getting the distance between the Mouse and the center of evilCircle with hypotenuse, continuing moving if hypotenuse is bigger than the radius.